tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-timing-function_two_values.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-timing-function, two values</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animation-timing-function_tag">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-timing-function can take two
   comma-separated values">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move-right {
         from {
            left: 0px;
         }
         to {
            left: 200px;
         }
      }
      @keyframes move-down {
         from {
            top: 0px;
         }
         to {
            top: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation-name: move-right, move-down;
         animation-duration: 1s;
         animation-fill-mode: forwards;
         animation-timing-function: steps(2), step-end;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         setTimeout(function() {
            passed = (style.left == '100px' && style.top == '0px');
         }, 750);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            passed = (passed && style.left == '200px' && style.top == '100px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should jump rightwards, then diagonally
   rightwards+downwards.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
